Erfolgreich JSON einfache GET Anfrage
