+++ app/src/main/java/de/info3/wegfinder24/WartebildschirmActivity.java
@@ -8,24 +8,17 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 
-import de.info3.wegfinder24.newtwork.Result;
-import de.info3.wegfinder24.newtwork.TestJSON.Summary;
+import de.info3.wegfinder24.newtwork.JSON.Example;
 import de.info3.wegfinder24.newtwork.OpenrouteService;
-import java.util.ArrayList;
 
-import de.info3.wegfinder24.newtwork.OpenrouteService;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 import retrofit2.Retrofit;
 import retrofit2.converter.gson.GsonConverterFactory;
 
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.ClientBuilder;
-import javax.ws.rs.client.Entity;
 //import javax.ws.rs.core.Response;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MediaType;
+
 //import javax.xml.transform.Result;
 
 public class WartebildschirmActivity extends AppCompatActivity {
@@ -60,49 +53,28 @@ public class WartebildschirmActivity extends AppCompatActivity {
 
                 OpenrouteService service = retrofit.create(OpenrouteService.class);
 
-                Call<Result> resultCall = service.listAuto1();
+                Call<Example> resultCall = service.listAuto1();
 
-                resultCall.enqueue(new Callback<Result>() {
+                resultCall.enqueue(new Callback<Example>() {
                     @Override
-                    public void onResponse(Call<Result> call, Response<Result> response) {
-                        Result result = response.body();
+                    public void onResponse(Call<Example> call, Response<Example> response) {
+                        Example example = response.body();
                         //int i = 0;
                         Log.d("FeaturesList","Anfrage ging durch");
                         if (response.code() != 200){
                             Log.d("FeaturesList","Problem, viel Glück!");
                             return;
                         }
-                        //Log.i("WartebildschirmActivity", String.valueOf(result.getfeaturesList().size()));
-                        //Log.d("Name"+' ' +i, result.getDepartureList().get(i).getServingLine().getName());
-                        //Log.d("Dauer",result.getfeaturesList().get(0).getzero().getproperties().getsummary().getduration());
-                        //TODO:de.info3.wegfinder24.newtwork.F_Null.getproperties()' on a null object reference kommt als Fehlermeldung
-                        //Log.d("Dauer", result.getfeaturesList().get(0).getzero().getproperties().getsummary().getduration().toString());
-                        //Log.d("Dauer",result.getFeatures().get(0).getProperties().getSummary().getDuration().toString());
-                        //Double Dauer = result.getFeatures().get(0).getProperties().getSummary().getDuration();
-                        //Double Dauer = Summary.
-                        //getFeatures().get(0).getProperties().getSummary().getDuration();
-                        //Log.d("Dauer", result.);
-                        //Intent intent = new Intent(WartebildschirmActivity.this, VarianteActivity.class);
-                        //intent.putExtra("Message",Dauer);
-                        //startActivity(intent);
 
-                        //Log.i("DepartureActivity", String.valueOf(result.getDepartureList().size()));
+                        Double Distanz = example.getFeatures().get(0).getProperties().getSummary().getDistance();
+                        Double Dauer = example.getFeatures().get(0).getProperties().getSummary().getDuration();
+                        Log.i("Distanz", Double.toString(Distanz));
+                        Log.i("Dauer", Double.toString(Dauer));
 
-                        for ( int i = 0; i < 40; i++)
-                        {
-                            //Log.i("Ausgabe", String.valueOf(result.getDepartureList().get(i)));
-                            //Log.i("Ausgabe" +i, result.getDepartureList().get(i).getServingLine().getName());
-                            //Log.i("Ausgabe" +i, result.getDepartureList().get(i).getServingLine().getNumber());
-                            //Log.i("Ausgabe" +i, result.getDepartureList().get(i).getServingLine().getDestination());
-                            Double Distanz = result.getDepartureList().get(i).getdidu().getDistance();
-                            Double Dauer = result.getDepartureList().get(i).getdidu().getDuration();
-                            Log.i("Ausgabe" +i, Double.toString(Distanz));
-                            Log.i("Ausgabe" +i, Double.toString(Dauer));
-                        }
                     }
 
                     @Override
-                    public void onFailure(Call<Result> call, Throwable t) {
+                    public void onFailure(Call<Example> call, Throwable t) {
                         Log.d("WartebildschirmActivity", "Anfragefehler");
                         //Intent intent = new Intent(WartebildschirmActivity.this, VarianteActivity.class);
                         //startActivity(intent);
