--- app/src/main/java/de/info3/wegfinder24/newtwork/DiDu/F_N_P_Summary.java
@@ -1,28 +1,0 @@
-package de.info3.wegfinder24.newtwork.DiDu;
-
-import com.google.gson.annotations.SerializedName;
-
-public class F_N_P_Summary {
-    /*@SerializedName("distance")
-    private Double distance;
-    public Double getdistance() {return this.distance;}
-    public void setdistance(Double distance) {this.distance = distance;}
-
-
-    @SerializedName("duration")
-    private Double duration;
-    public Double getduration() {return this.duration;}
-    public void setduration(Double duration) {this.duration = duration;}
-    */
-
-    @SerializedName("distance")
-    private Double distance;
-    public Double getDistance() {return distance;}
-    public void setDistance(Double distance) {this.distance = distance;}
-
-    @SerializedName("duration")
-    private Integer duration;
-    public Integer getDuration() {return duration;}
-    public void setDuration(Integer duration) {this.duration = duration;}
-
-}
