--- app/src/main/java/de/info3/wegfinder24/newtwork/DiDu/F_N_Properties.java
@@ -1,15 +1,0 @@
-package de.info3.wegfinder24.newtwork.DiDu;
-
-import com.google.gson.annotations.SerializedName;
-
-public class F_N_Properties {
-    @SerializedName("summary")
-    /*private F_N_P_Summary summary;
-    public F_N_P_Summary getsummary() {return this.summary; }
-    public void setsummary(F_N_P_Summary summary) {this.summary = summary;}
-    */
-    private F_N_P_Summary summary;
-    public F_N_P_Summary getSummary() {return summary;}
-    public void setSummary(F_N_P_Summary summary) {this.summary = summary;}
-
-}
