--- app/src/main/java/de/info3/wegfinder24/newtwork/F_Null.java
@@ -1,14 +1,0 @@
-package de.info3.wegfinder24.newtwork;
-
-import com.google.gson.annotations.SerializedName;
-
-import de.info3.wegfinder24.newtwork.DiDu.F_N_Properties;
-
-public class F_Null {
-    /*@SerializedName("properties")
-    private F_N_Properties properties;
-
-    public F_N_Properties getproperties() {return this.properties;}
-
-    public void setproperties(F_N_Properties properties) {this.properties = properties;}*/
-}
